# bls-server/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "BLS-KOX API",
    "version": "1.0.0",
    "description": "BLS-KOX 多租户后台管理系统 API 文档\n\n## 认证\n在请求头添加 `Authorization: Bearer <token>`\n登录接口获取 token：`POST /api/auth/login`\n\n## 默认账号\n- 账号: `superadmin`  密码: `123456`  租户: `000000`\n- 账号: `admin`       密码: `123456`  租户: `100000`",
    "contact": {
      "name": "BLS-KOX",
      "url": "https://github.com/npcxl/BLS-KOX"
    }
  },
  "servers": [
    {
      "url": "http://localhost:6001",
      "description": "本地开发"
    },
    {
      "url": "/api",
      "description": "Nginx 反代（生产）"
    }
  ],
  "paths": {
...
    "/api/system/storage/{storageId}": {
      "get": {
        "tags": [
          "storage"
        ],
        "summary": "storage - 查询 :storageId",
        "operationId": "get_api_system_storage_storageId_",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "操作成功"
                    },
                    "data": {
                      "type": "object",
                      "description": "响应数据"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "参数错误"
          }
        },
        "parameters": [
          {
            "name": "tenantId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "tenantId"
          },
          {
            "name": "storageId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "storageId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/storage_get__api_system_storage_storageId_"
              }
            }
          }
        }
      }
    },
    "/api/system/storage/add": {
      "post": {
        "tags": [
          "storage"
        ],
        "summary": "storage - 新增记录",
        "operationId": "post_api_system_storage_add",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "操作成功"
                    },
                    "data": {
                      "type": "object",
                      "description": "响应数据"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "参数错误"
          }
        },
        "parameters": [
          {
            "name": "tenantId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "tenantId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/storage_post__api_system_storage_add"
              }
            }
          }
        }
      }
    },
    "/api/system/storage/edit": {
      "put": {
        "tags": [
          "storage"
        ],
        "summary": "storage - 修改记录",
        "operationId": "put_api_system_storage_edit",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "操作成功"
                    },
                    "data": {
                      "type": "object",
                      "description": "响应数据"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "参数错误"
          }
        },
        "parameters": [
          {
            "name": "tenantId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "tenantId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/storage_put__api_system_storage_edit"
              }
            }
          }
        }
      }
    },
    "/api/system/storage/file/{fileId}": {
      "delete": {
        "tags": [
          "storage"
        ],
        "summary": "storage - 删除 :fileId",
        "operationId": "delete_api_system_storage_file_fileId_",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "操作成功"
                    },
                    "data": {
                      "type": "object",
                      "description": "响应数据"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "参数错误"
          }
        },
        "parameters": [
          {
            "name": "tenantId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "tenantId"
          },
          {
            "name": "fileId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "fileId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/storage_delete__api_system_storage_file_fileId_"
              }
            }
          }
        }
      }
    },
    "/api/system/storage/file/{fileId}/download": {
      "get": {
        "tags": [
          "storage"
        ],
        "summary": "storage - 查询 download",
        "operationId": "get_api_system_storage_file_fileId_download",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "操作成功"
                    },
                    "data": {
                      "type": "object",
                      "description": "响应数据"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "参数错误"
          }
        },
        "parameters": [
          {
            "name": "tenantId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "tenantId"
          },
          {
            "name": "fileId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "fileId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/storage_get__api_system_storage_file_fileId_download"
              }
            }
          }
        }
      }
    },
    "/api/system/storage/file/{fileId}/url": {
      "get": {
        "tags": [
          "storage"
        ],
        "summary": "storage - 查询 url",
        "operationId": "get_api_system_storage_file_fileId_url",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "操作成功"
                    },
                    "data": {
                      "type": "object",
                      "description": "响应数据"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "参数错误"
          }
        },
        "parameters": [
          {
            "name": "tenantId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "tenantId"
          },
          {
            "name": "fileId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "fileId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/storage_get__api_system_storage_file_fileId_url"
              }
            }
          }
        }
      }
    },
    "/api/system/storage/files": {
      "get": {
        "tags": [
          "storage"
        ],
        "summary": "storage - 查询 files",
        "operationId": "get_api_system_storage_files",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "操作成功"
                    },
                    "data": {
                      "type": "object",
                      "description": "响应数据"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "参数错误"
          }
        },
        "parameters": [
          {
            "name": "tenantId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "tenantId"
          },
          {
            "name": "pageNum",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "pageNum"
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "pageSize"
          },
          {
            "name": "originalName",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "originalName"
          },
          {
            "name": "moduleName",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "moduleName"
          },
          {
            "name": "accessType",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "accessType"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/storage_get__api_system_storage_files"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "storage_get__api_system_storage_storageId_": {
        "type": "object",
        "properties": {
          "tenantId": {
            "type": "string",
            "description": "tenantId"
          }
        }
      },
      "storage_post__api_system_storage_add": {
        "type": "object",
        "properties": {
          "tenantId": {
            "type": "string",
            "description": "tenantId"
          }
        }
      },
      "storage_put__api_system_storage_edit": {
        "type": "object",
        "properties": {
          "tenantId": {
            "type": "string",
            "description": "tenantId"
          }
        }
      },
      "storage_delete__api_system_storage_file_fileId_": {
        "type": "object",
        "properties": {
          "tenantId": {
            "type": "string",
            "description": "tenantId"
          }
        }
      },
      "storage_get__api_system_storage_file_fileId_download": {
        "type": "object",
        "properties": {
          "tenantId": {
            "type": "string",
            "description": "tenantId"
          }
        }
      },
      "storage_get__api_system_storage_file_fileId_url": {
        "type": "object",
        "properties": {
          "tenantId": {
            "type": "string",
            "description": "tenantId"
          }
        }
      },
      "storage_get__api_system_storage_files": {
        "type": "object",
        "properties": {
          "tenantId": {
            "type": "string",
            "description": "tenantId"
          }
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "登录获取的 accessToken"
      }
    }
  },
...
}